Refactorings | API Performance - Network bandwidth reduction| Junits updated

+++ src/main/java/org/project/currencyconverter/model/ExchangeRates.java
@@ -2,6 +2,7 @@ package org.project.currencyconverter.model;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import java.io.Serializable;
+import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -18,7 +19,7 @@ public class ExchangeRates implements Serializable
 {
     private static final long serialVersionUID = -5932239093129460014L;
     private String targetCurrency;
-    private double exchangeRate;
+    private BigDecimal exchangeRate;
 
 
     @Override
